chore(snuba): Move resolve datetime64 to shared utils
