+++ tests/sentry/search/events/test_fields.py
@@ -1,3 +1,5 @@
+from datetime import UTC, datetime
+
 import pytest
 from snuba_sdk.column import Column
 from snuba_sdk.function import Function
@@ -11,6 +13,7 @@ from sentry.search.events.fields import (
     parse_arguments,
     parse_combinator,
     parse_function,
+    resolve_datetime64,
 )
 from sentry.snuba.dataset import Dataset
 
@@ -285,3 +288,45 @@ def test_combinator_names_are_reserved(combinator):
         assert not function.endswith(
             combinator.kind
         ), f"Cannot name function `{function}` because `-{combinator.kind}` suffix is reserved for combinators"
+
+
+@pytest.mark.parametrize(
+    ["value", "expected"],
+    [
+        pytest.param(None, None, id="None"),
+        pytest.param(
+            1721339541.429418,
+            Function("toDateTime64", [1721339541.429418, 6]),
+            id="1721339541.429418",
+        ),
+        pytest.param(
+            datetime(
+                year=2024,
+                month=7,
+                day=18,
+                hour=17,
+                minute=55,
+                second=27,
+                microsecond=72215,
+            ),
+            Function("toDateTime64", ["2024-07-18T17:55:27.072215", 6]),
+            id="2024-07-18T17:55:27.072215",
+        ),
+        pytest.param(
+            datetime(
+                year=2024,
+                month=7,
+                day=18,
+                hour=17,
+                minute=55,
+                second=27,
+                microsecond=72215,
+                tzinfo=UTC,
+            ),
+            Function("toDateTime64", ["2024-07-18T17:55:27.072215", 6]),
+            id="2024-07-18T17:55:27.072215 with timezone",
+        ),
+    ],
+)
+def test_resolve_datetime64(value, expected):
+    assert resolve_datetime64(value) == expected
